+++ lending/RELEASES.md
@@ -219,3 +219,13 @@ Based on:
 - [python v11.0.2] lending
 ### Releases
 - [PyPI v11.0.2] https://pypi.org/project/codat-lending/11.0.2 - lending
+
+## 2026-08-28 08:59:53
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- codat sdk-codegen db72283147925427108432b8413c2487af21f9e6 https://github.com/codat-spikes/sdk-codegen
+### Generated
+- [python v11.0.3] lending
+### Releases
+- [PyPI v11.0.3] https://pypi.org/project/codat-lending/11.0.3 - lending